Start work on rewrite
