+++ docs/api/rule_context.md
@@ -1,0 +1,4 @@
+# Rule context
+
+When executed, Transdoc rules can access a `TransdocContext` object to get
+information about the current transformation.
