+++ docs/builtins.md
@@ -1,0 +1,5 @@
+# Built-in Transdoc rules
+
+Transdoc includes a number of common rules.
+
+
