+++ transdoc/handlers/__init__.py
@@ -1,0 +1,9 @@
+"""
+# Transdoc / Handlers
+
+Code defining Transdoc handlers.
+"""
+from .api import TransdocHandler
+
+
+__all__ = ['TransdocHandler']
